requirements.txt

greenlet==3.3.2
numpy==2.4.4
pandas==3.0.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0

